* changes according to suggestions
